FIX 합치지 쉽도록 구조 변경
